# opt: Change Memo data structures