--- a/pkg/sql/opt/exec/execbuilder/testdata/select_index
+++ b/pkg/sql/opt/exec/execbuilder/testdata/select_index
@@ -928,8 +928,8 @@ EXPLAIN (VERBOSE) SELECT x FROM xy WHERE y > 0 AND y < 2 ORDER BY y
 ----
 render           ·         ·            (x)                 ·
  │               render 0  x            ·                   ·
- └── index-join  ·         ·            (x, y)              ·
-      ├── scan   ·         ·            (y, rowid[hidden])  ·
+ └── index-join  ·         ·            (x, y)              +y
+      ├── scan   ·         ·            (y, rowid[hidden])  +y
       │          table     xy@xy_y_idx  ·                   ·
       │          spans     /1-/2        ·                   ·
       └── scan   ·         ·            (x, y)              ·