… pkg/sql/opt/bench/bench_test.go
… pkg/sql/opt/exec/execbuilder/relational_builder.go
… pkg/sql/opt/idxconstraint/index_constraints.go
… pkg/sql/opt/idxconstraint/index_constraints_test.go
… pkg/sql/opt/memo/best_expr.go
… pkg/sql/opt/memo/check_expr.go
… pkg/sql/opt/memo/constraint_builder.go
… pkg/sql/opt/memo/dynamic.go
… pkg/sql/opt/memo/dynamic_test.go
… pkg/sql/opt/memo/expr.go
… pkg/sql/opt/memo/expr_format.go
… pkg/sql/opt/memo/expr_view_format.go
… pkg/sql/opt/memo/expr_view_test.go
… pkg/sql/opt/memo/extract.go
… pkg/sql/opt/memo/group.go
… pkg/sql/opt/memo/interner.go
… pkg/sql/opt/memo/interner_test.go
… pkg/sql/opt/memo/list_storage.go
… pkg/sql/opt/memo/list_storage_test.go
… pkg/sql/opt/memo/logical_props_builder.go
… pkg/sql/opt/memo/logical_props_builder_test.go
… pkg/sql/opt/memo/memo.go
… pkg/sql/opt/memo/memo_format.go
… pkg/sql/opt/memo/memo_test.go
… pkg/sql/opt/memo/private_defs.go
… pkg/sql/opt/memo/private_defs_test.go
… pkg/sql/opt/memo/private_storage.go
… pkg/sql/opt/memo/private_storage_test.go
… pkg/sql/opt/memo/statistics_builder.go
… pkg/sql/opt/memo/statistics_builder_test.go