Change how the Memo stores groups and expressions. Currently, both
are stored in Go slices which are resized as new groups and expressions
are added to the memo. Expressions store GroupIDs and PrivateIDs, which
are indexes into slices holding memo groups and expression-specific  data.

While this design has worked OK, it has several downsides:

  1. There is very little type safety, since all expressions are stored
     in identical 16 byte structs so they can be packed into slices.
     This can also make debugging tedious, as the debugger does not
     know how to decode the expressions.

  2. Expressions can have only 3 fields, since we pack them into 16
     bytes. For many operators this is fine, but it requires awkward
     workarounds for operators than need >3 fields.

  3. Many nodes have "private" data fields that require the allocation
     of a separate object to store. This cancels what would otherwise
     be an advantage of the current design, since it ends up allocating
     about the same number of objects as the new design for this and
     other reasons. It also requires tedious calls to InternPrivate
     and LookupPrivate methods in order to map to and from slice
     indexes.

  4. When the optimizer decides on a lowest cost expression tree, it
     must store that information "off to the side" in BestExpr
     structures so that ExprView knows which of the trees in the forest
     to traverse. This triggers additional allocations.

  5. Traversal of the tree requires an auxiliary ExprView data structure
     that holds onto context necessary for the traversal (i.e. the Memo).
     This makes the current design harder to understand and use.

  6. All expressions allocate and store logical properties. While this is
     necessary for relational expressions, it is not required for most
     scalar operators.

  7. The memo expression storage format is not compatible with nodes used
     in other parts of Cockroach, like AST and planNodes. This makes mixing
     and matching nodes difficult, and requires "wrapper" nodes to map
     between different representations.

The new memo expression representation solves these problems using a more
natural Go-friendly design. Each expression is a differently-typed Go
struct that implements the Expr interface:
```
  type Expr interface {
    Op() Operator
    ChildCount() int
    Child(nth int) Expr
    Private() interface{}
    String() string
  }
```
This is similar to how AST nodes and planNodes work. Optgen generates a
struct definition for each operator, as well as implementations for the
Expr interface methods. Private values are stored inline in the structs,
replacing the need for an additional allocation. Expressions can be
constructed using a very natural factory interface with no explicit
interning required.

Where it gets more interesting is for relational expressions, which are
part of memo groups that can contain multiple logically equivalent
expressions. The new representation links together expressions in the
same group using "next" pointers. In addition, the first expression in
the memo is allocated with extra inline storage to store relational and
physical properties. This avoids separate allocation for the properties
like the current representation requires. For a simple expression tree
with 1 expression per group, the tree requires just 1 object per memo
group.

Here are numbers for the old vs. new representations:
```
name                               old time/op  new time/op  delta
Phases/kv-read/ExecBuild           15.5µs ± 3%  10.7µs ± 3%  -31.03%  (p=0.008 n=5+5)
Phases/kv-read-no-prep/ExecBuild   38.1µs ± 3%  33.4µs ± 2%  -12.14%  (p=0.008 n=5+5)
Phases/kv-read-const/ExecBuild      682ns ± 2%   622ns ± 2%   -8.71%  (p=0.008 n=5+5)
Phases/tpcc-new-order/ExecBuild    53.2µs ± 5%  35.4µs ± 3%  -33.41%  (p=0.008 n=5+5)
Phases/tpcc-delivery/ExecBuild     38.9µs ± 3%  23.8µs ± 5%  -38.72%  (p=0.008 n=5+5)
Phases/tpcc-stock-level/ExecBuild   167µs ± 2%   114µs ± 3%  -31.34%  (p=0.008 n=5+5)
Exec/kv-read/ExecPlan               197µs ± 2%   193µs ± 2%     ~     (p=0.056 n=5+5)
Exec/kv-read/ExecOpt                186µs ± 4%   189µs ± 3%     ~     (p=0.421 n=5+5)
Exec/kv-read-no-prep/ExecPlan       318µs ± 3%   315µs ± 2%     ~     (p=0.548 n=5+5)
Exec/kv-read-no-prep/ExecOpt        340µs ± 8%   327µs ± 1%   -4.00%  (p=0.016 n=5+5)
Exec/kv-read-const/ExecPlan         194µs ± 3%   192µs ± 2%     ~     (p=0.421 n=5+5)
Exec/kv-read-const/ExecOpt          161µs ± 3%   161µs ± 4%     ~     (p=1.000 n=5+5)
Exec/tpcc-new-order/ExecPlan        301µs ± 8%   268µs ± 2%  -10.81%  (p=0.008 n=5+5)
Exec/tpcc-new-order/ExecOpt         269µs ± 1%   261µs ± 3%   -3.09%  (p=0.032 n=5+5)
Exec/tpcc-delivery/ExecPlan         243µs ± 2%   253µs ± 9%     ~     (p=0.690 n=5+5)
Exec/tpcc-delivery/ExecOpt          242µs ± 1%   245µs ± 6%     ~     (p=0.690 n=5+5)
Exec/tpcc-stock-level/ExecPlan      545µs ± 2%   525µs ± 2%   -3.53%  (p=0.008 n=5+5)
Exec/tpcc-stock-level/ExecOpt       532µs ± 3%   501µs ± 3%   -5.79%  (p=0.008 n=5+5)

name                               old alloc/op   new alloc/op   delta

Phases/kv-read/ExecBuild             5.39kB ± 0%    3.90kB ± 0%  -27.53%  (p=0.008 n=5+5)

Phases/kv-read-no-prep/ExecBuild     10.3kB ± 0%     9.2kB ± 0%  -10.69%  (p=0.008 n=5+5)

Phases/kv-read-const/ExecBuild         264B ± 0%      216B ± 0%  -18.18%  (p=0.008 n=5+5)

Phases/tpcc-new-order/ExecBuild      14.9kB ± 0%     9.9kB ± 0%  -33.79%  (p=0.008 n=5+5)

Phases/tpcc-delivery/ExecBuild       11.5kB ± 0%     6.4kB ± 0%  -43.83%  (p=0.000 n=5+4)

Phases/tpcc-stock-level/ExecBuild    50.1kB ± 0%    30.1kB ± 0%  -40.08%  (p=0.008 n=5+5)

Exec/kv-read/ExecPlan                30.3kB ± 0%    26.7kB ± 0%  -11.77%  (p=0.029 n=4+4)

Exec/kv-read/ExecOpt                 26.4kB ± 0%    24.9kB ± 0%   -5.65%  (p=0.008 n=5+5)

Exec/kv-read-no-prep/ExecPlan        43.3kB ± 2%    39.3kB ± 0%   -9.19%  (p=0.016 n=5+4)

Exec/kv-read-no-prep/ExecOpt         44.5kB ± 0%    39.8kB ± 0%  -10.50%  (p=0.029 n=4+4)

Exec/kv-read-const/ExecPlan          29.9kB ± 0%    26.3kB ± 0%  -12.02%  (p=0.029 n=4+4)

Exec/kv-read-const/ExecOpt           21.0kB ± 0%    20.9kB ± 0%   -0.24%  (p=0.029 n=4+4)

Exec/tpcc-new-order/ExecPlan         66.3kB ± 0%    57.9kB ± 0%  -12.68%  (p=0.016 n=4+5)

Exec/tpcc-new-order/ExecOpt          51.3kB ± 0%    46.3kB ± 0%   -9.88%  (p=0.029 n=4+4)

Exec/tpcc-delivery/ExecPlan          33.9kB ± 0%    30.1kB ± 0%  -11.13%  (p=0.016 n=5+4)

Exec/tpcc-delivery/ExecOpt           35.7kB ± 0%    30.7kB ± 0%  -14.07%  (p=0.016 n=4+5)

Exec/tpcc-stock-level/ExecPlan        150kB ± 0%     138kB ± 0%   -7.57%  (p=0.029 n=4+4)

Exec/tpcc-stock-level/ExecOpt         131kB ± 0%     111kB ± 0%  -15.31%  (p=0.016 n=5+4)

name                               old allocs/op  new allocs/op  delta

Phases/kv-read/ExecBuild               34.0 ± 0%      29.0 ± 0%  -14.71%  (p=0.008 n=5+5)

Phases/kv-read-no-prep/ExecBuild       96.0 ± 0%      87.0 ± 0%   -9.38%  (p=0.008 n=5+5)

Phases/kv-read-const/ExecBuild         6.00 ± 0%      5.00 ± 0%  -16.67%  (p=0.008 n=5+5)

Phases/tpcc-new-order/ExecBuild        76.0 ± 0%      57.0 ± 0%  -25.00%  (p=0.008 n=5+5)

Phases/tpcc-delivery/ExecBuild         62.0 ± 0%      43.0 ± 0%  -30.65%  (p=0.008 n=5+5)

Phases/tpcc-stock-level/ExecBuild       217 ± 0%       171 ± 0%  -21.20%  (p=0.008 n=5+5)

Exec/kv-read/ExecPlan                   315 ± 0%       310 ± 0%   -1.59%  (p=0.029 n=4+4)

Exec/kv-read/ExecOpt                    273 ± 0%       267 ± 0%   -2.05%  (p=0.008 n=5+5)

Exec/kv-read-no-prep/ExecPlan           460 ± 0%       456 ± 0%   -0.96%  (p=0.016 n=4+5)

Exec/kv-read-no-prep/ExecOpt            438 ± 0%       413 ± 0%   -5.71%  (p=0.029 n=4+4)

Exec/kv-read-const/ExecPlan             308 ± 0%       303 ± 0%   -1.62%  (p=0.029 n=4+4)

Exec/kv-read-const/ExecOpt              238 ± 0%       237 ± 0%   -0.42%  (p=0.029 n=4+4)

Exec/tpcc-new-order/ExecPlan            414 ± 0%       401 ± 0%     ~     (p=0.079 n=4+5)

Exec/tpcc-new-order/ExecOpt             350 ± 0%       331 ± 0%   -5.43%  (p=0.029 n=4+4)

Exec/tpcc-delivery/ExecPlan             363 ± 0%       354 ± 0%   -2.48%  (p=0.000 n=5+4)

Exec/tpcc-delivery/ExecOpt              326 ± 0%       307 ± 0%     ~     (p=0.079 n=4+5)

Exec/tpcc-stock-level/ExecPlan        1.04k ± 0%     1.01k ± 0%   -2.22%  (p=0.029 n=4+4)

Exec/tpcc-stock-level/ExecOpt           763 ± 0%       717 ± 0%   -6.03%  (p=0.000 n=5+4)
```