@RaduBerinde 2018-10-12
Hi Andy,

I put an initial set of comments for the `memo` part here, as a diff (comments usually refer to the line below): https://gist.github.com/RaduBerinde/a3d238f44495b1822bc88fb77f6b1880

Feel free to comment in that gist, I think that's easier than interleaving all discussions in this PR.