@RaduBerinde 2018-10-15
___
*[pkg/sql/opt/optbuilder/testdata/aggregate, line 894 at r15](https://reviewable.io/reviews/cockroachdb/cockroach/31173#-LOY_FuH5Ra9aJY65fpb:-LOsdDNY2s_I3QAme6L-:bk9uz74) ([raw file](https://github.com/cockroachdb/cockroach/blob/9858934f2e50787a5984d308a63189cf0392f7f2/pkg/sql/opt/optbuilder/testdata/aggregate#L894)):*
<details><summary><i>Previously, andy-kimball (Andy Kimball) wrote…</i></summary><blockquote>

Yeah, I went back and forth on this. I could make it `true`, though it might lead to other confusion, since people might then think there's actually a `true` expression when there isn't. @radu, @rebecca, any opinion on this?
</blockquote></details>

I would make it show up as `no-filters` (or `filters (none)` or `filters (true)`)


<!-- Sent from Reviewable.io -->