@andy-kimball 2018-10-15
> I would make it show up as `no-filters` (or `filters (none)` or `filters (true)`)

Good idea, I like `filters (true)`.