@RaduBerinde 2018-10-18
The updates to memo and xform LGTM.

One small thing (feel free to take it or leave it) -  in the interner it would be cleaner if `Next` would only do the "read-only" stuff, and `Add` would do all the modifications; something along the lines of: https://gist.github.com/RaduBerinde/e10a1b2b5393b9e09b0494c714a964f7